requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Google Cloud dependencies
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.17.0

# Firebase
firebase-admin==6.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Caching
redis==5.0.1
cachetools==5.3.2

# Rate limiting
slowapi==0.1.9

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# Development and testing (optional, can be removed in production)
pytest==7.4.3
pytest-asyncio==0.21.1
fakeredis==2.20.1

# Performance
orjson==3.9.10
python-multipart==0.0.6 